[frontend] renaming threatActor into threatActorGroup (#issue/3166)
